+++ EXAMPLES.md
@@ -1,0 +1,11 @@
+# Esimerkkitiedostoja
+
+![Build and deploy](https://github.com/jaantaponen/verotunkki/actions/workflows/build-deploy.yml/badge.svg)
+
+* [Coinbase](https://github.com/jaantaponen/verotunkki/blob/update-terms/tests/files/SortIssuedCoinbase.csv)
+* [CoinbasePro](https://github.com/jaantaponen/verotunkki/blob/update-terms/tests/files/SortIssuedCoinbasePro.csv)
+* [Degiro](https://github.com/jaantaponen/verotunkki/blob/update-terms/tests/files/transactionsDegiro.csv)
+
+### Lisää esimerkkejä?
+
+Mikäli haluat tarkastella testien pohjatiedostoja ne löytyvät [täältä](https://github.com/jaantaponen/verotunkki/tree/update-terms/tests/files)
