+++ package.json
@@ -4,30 +4,30 @@
   "private": true,
   "homepage": "https://verotunkki.fi",
   "dependencies": {
-    "@emotion/react": "^11.7.0",
-    "@emotion/styled": "^11.6.0",
-    "@mui/icons-material": "^5.2.1",
-    "@mui/material": "^5.2.3",
-    "@mui/types": "^7.1.0",
-    "@mui/x-data-grid": "^5.2.1",
+    "@emotion/react": "^11.8.2",
+    "@emotion/styled": "^11.8.1",
+    "@mui/icons-material": "^5.5.1",
+    "@mui/material": "^5.5.3",
+    "@mui/types": "^7.1.3",
+    "@mui/x-data-grid": "^5.7.0",
     "@testing-library/jest-dom": "^5.16.1",
     "@testing-library/user-event": "^12.8.3",
     "@types/node": "^12.20.37",
-    "@types/react": "^17.0.37",
+    "@types/react": "^17.0.43",
     "@types/react-dom": "^17.0.11",
-    "axios": "^0.24.0",
-    "csv": "^6.0.4",
+    "axios": "^0.26.1",
+    "csv": "^6.0.5",
     "lodash": "4.17.21",
     "moment": "^2.29.1",
     "moment-timezone": "^0.5.34",
-    "nanoid": "^3.1.30",
+    "nanoid": "^3.3.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "react-mui-dropzone": "^4.0.5",
-    "react-router-dom": "^6.1.1",
+    "react-mui-dropzone": "^4.0.6",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "4.0.3",
-    "ts-node": "^10.4.0",
-    "web-vitals": "^1.1.2"
+    "ts-node": "^10.7.0",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
